@@ -19,7 +19,7 @@
         "symfony/css-selector": "^5.1",
         "symfony/dotenv": "5.1.*",
         "symfony/expression-language": "5.1.*",
-        "symfony/flex": "^1.3.1",
+        "symfony/flex": "^1.9",
         "symfony/form": "5.1.*",
         "symfony/framework-bundle": "5.1.*",
         "symfony/http-client": "5.1.*",
@@ -45,6 +45,7 @@
         "vich/uploader-bundle": "^1.15"
     },
     "require-dev": {
+        "doctrine/doctrine-fixtures-bundle": "^3.3",
         "symfony/debug-bundle": "^5.1",
         "symfony/maker-bundle": "^1.0",
         "symfony/stopwatch": "^5.1",
